+++ packages/@expo/cli/bin/cli.ts
@@ -39,6 +39,7 @@ const commands: { [command: string]: () => Promise<Command> } = {
   run: () => import('../src/run/index.js').then((i) => i.expoRun),
   'run:ios': () => import('../src/run/ios/index.js').then((i) => i.expoRunIos),
   'run:android': () => import('../src/run/android/index.js').then((i) => i.expoRunAndroid),
+  'run:macos': () => import('../src/run/macos/index.js').then((i) => i.expoRunMacos),
   start: () => import('../src/start/index.js').then((i) => i.expoStart),
   prebuild: () => import('../src/prebuild/index.js').then((i) => i.expoPrebuild),
   config: () => import('../src/config/index.js').then((i) => i.expoConfig),